Refactor node_flasher to use more sane logic
The infinite chaining of callback functions made it impossible to maintain a reference to the tween that was being called. Instead, within node_flash we use set_loops() on a single tween to make it loop infinitely. We also return the infinitely looping tween so that we can get a reference to it.